Refactoring and making the style selection use int instead of bool for more stylesheet options
