+++ nicos_ess/gui/widgets/utils.py
@@ -1,0 +1,25 @@
+from nicos.guisupport.qt import pyqtProperty
+
+
+class StyleSelector:
+
+    def __init__(self):
+        self._style_type = 0
+
+    @pyqtProperty(int)
+    def style_type(self):
+        return self._style_type
+
+    @style_type.setter
+    def style_type(self, value):
+        self._style_type = value
+
+    def refresh_gui(self):
+        try:
+            self.style().unpolish(self)
+            self.style().polish(self)
+            self.update()
+        except AttributeError as e:
+            print(e)
+
+
