+++ nicos_ess/guiconfig.qss
@@ -525,11 +525,20 @@ QMessageBox > QLabel {
 
 /* ----------------------------------------- */
 
-CommandLineEdit {
-font-size: 20;
-min-height: 22px;
-border: 2px solid #0094ca;
-border-radius: 5px;
+CommandLineEditStyleSheet[style_type="0"] {
+    font-size: 20;
+    min-height: 22px;
+    border: 2px solid #0094ca;
+    border-radius: 5px;
+    background-color: white;
+}
+
+CommandLineEditStyleSheet[style_type="1"] {
+    font-size: 20;
+    min-height: 22px;
+    border: 2px solid #0094ca;
+    border-radius: 5px;
+    background-color: #ffdddd;
 }
 
 Panel[panel='ExpPanel'] QLineEdit,
