+++ nicos_ess/loki/guiconfig.qss
@@ -505,19 +505,20 @@ QMessageBox > QLabel {
 
 /* ----------------------------------------- */
 
-CommandLineEditStyleSheet[change_style="true"] {
-    font-size: 20;
-    min-height: 22px;
-    border: 2px solid #0094ca;
-    border-radius: 5px;
-    background-color: #ffdddd;
-}
-CommandLineEditStyleSheet[change_style="false"] {
+CommandLineEditStyleSheet[style_type="0"] {
     font-size: 20;
     min-height: 22px;
     border: 2px solid #0094ca;
     border-radius: 5px;
     background-color: white;
+}
+
+CommandLineEditStyleSheet[style_type="1"] {
+    font-size: 20;
+    min-height: 22px;
+    border: 2px solid #0094ca;
+    border-radius: 5px;
+    background-color: #ffdddd;
 }
 
 Panel[panel='ExpPanel'] QLineEdit,
